Add pagination
For Jekyll 3, you now are required to include the jekyll-paginate
plugin in your Gemfile and in your _config.yml under gems as described
in https://jekyllrb.com/docs/pagination/.

Note: I heard (https://teamtreehouse.com/community/jekyllpaginate-gem)
that jekyll-gist was also required in the _config.yml but I didn’t have
a problem so I left it out. Try to include it if there are issues.